freeze migrations into v0.15.8_9.ts, bump to 0.15.8:10
